Extract sass prefixer from rollup config
