+++ rollup.config.js
@@ -1,4 +1,5 @@
 import jsPrefixer, { componentInfo } from './rollup-plugin-javascript-prefixer';
+import sassPrefixer from './sass-prefixer';
 
 const nodeResolve = require('rollup-plugin-node-resolve');
 const commonjs = require('rollup-plugin-commonjs');
@@ -43,28 +44,8 @@ const commonPlugins = [
         .use(
           classPrefixer({
             prefix: componentInfo.prefix,
-            transform: (prefX, selector) => {
-              // Exclude tags, only apply to classes
-              const splitSelectors = selector.split('.');
-              const secondArray = [];
-              for (let i = 0; i < splitSelectors.length; ++i) {
-                if (
-                  splitSelectors[i].startsWith(
-                    componentInfo.standardComponentClassPrefix
-                  )
-                ) {
-                  secondArray.push(`.${prefX}_${splitSelectors[i]}`);
-                  // If index is null, that means there was no empty string
-                  // before, which means it was not a tag but a class and needs
-                  // to be labeled as such.
-                } else if (i !== 0) {
-                  secondArray.push(`.${splitSelectors[i]}`);
-                } else {
-                  secondArray.push(splitSelectors[i]);
-                }
-              }
-              return secondArray.join('');
-            },
+            transform: (prefX, selector) =>
+              sassPrefixer(prefX, selector, componentInfo),
           })
         )
         .process(css, { from: undefined })
